Move spirv_generator into _kernel_api_impl.spirv
